添加通发送邮箱

+++ service/service_education/src/main/java/com/sustech/service_education/entity/Course.java
@@ -19,5 +19,5 @@ public class Course {
 
     String status;
 
-    Float fee;
+    Double fee;
 }
